dags/weather_data_pipeline.py
from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python_operator import PythonOperator

def fetch_weather_data():
    # Function to fetch weather data from an API
    pass

def process_weather_data():
    # Function to process the fetched weather data
    pass

def store_weather_data():
    # Function to store the processed weather data
    pass

default_args = {
    'owner': 'airflow',
    'depends_on_past': False,
    'email_on_failure': False,
    'email_on_retry': False,
    'retries': 1,
    'retry_delay': timedelta(minutes=5),
}

dag = DAG(
    'weather_data_pipeline',
    default_args=default_args,
    description='A simple weather data pipeline',
    schedule_interval=timedelta(days=1),
    start_date=datetime(2023, 1, 1),
    catchup=False,
)

fetch_task = PythonOperator(
    task_id='fetch_weather_data',
    python_callable=fetch_weather_data,
    dag=dag,
)

process_task = PythonOperator(
    task_id='process_weather_data',
    python_callable=process_weather_data,
    dag=dag,
)

store_task = PythonOperator(
    task_id='store_weather_data',
    python_callable=store_weather_data,
    dag=dag,
)

fetch_task >> process_task >> store_task